# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Optrispace Server API",
        "title": "Optrispace API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/contracts": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Returns list of contracts where the current user is performer or customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contract"
                ],
                "summary": "List contracts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ContractDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "user not authorized",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/echo.HTTPError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creates a new contract based on existent application.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contract"
                ],
                "summary": "Create a new contract",
                "parameters": [
                    {
                        "description": "Contract Params",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.createContractParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.ContractDTO"
                        }
                    },
                    "400": {
                        "description": "invalid format",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "409": {
                        "description": "duplication",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/echo.HTTPError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/contracts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Returns contract with specified id. This operation is allowed only for performer or customer.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contract"
                ],
                "summary": "Get contract",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ContractDTO"
                        }
                    },
                    "401": {
                        "description": "user not authorized",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "404": {
                        "description": "contract not found or user not authorized to view contract",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/echo.HTTPError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/contracts/{id}/accept": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Performer is accepting contract",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contract"
                ],
                "summary": "Accept contract",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ContractDTO"
                        }
                    },
                    "401": {
                        "description": "user not authorized",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "403": {
                        "description": "insufficient rights",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "404": {
                        "description": "contract not found or user not authorized to view contract",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/echo.HTTPError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/contracts/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Customer is approving performer's working results",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contract"
                ],
                "summary": "Approve working results and allow to withdraw money from Smart Contract",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ContractDTO"
                        }
                    },
                    "401": {
                        "description": "user not authorized",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "403": {
                        "description": "insufficient rights",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "404": {
                        "description": "contract not found or user not authorized to view contract",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "422": {
                        "description": "insufficient funds on the contract in the blockchain",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/echo.HTTPError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/contracts/{id}/complete": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Performer is completing contract",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contract"
                ],
                "summary": "Complete contract",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ContractDTO"
                        }
                    },
                    "401": {
                        "description": "user not authorized",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "403": {
                        "description": "insufficient rights",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "404": {
                        "description": "contract not found or user not authorized to view contract",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/echo.HTTPError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/contracts/{id}/deploy": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Customer has deployed contract to the blockchain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contract"
                ],
                "summary": "Deploy contract",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ContractDTO"
                        }
                    },
                    "400": {
                        "description": "invalid format",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "401": {
                        "description": "user not authorized",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "403": {
                        "description": "insufficient rights",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "404": {
                        "description": "contract not found or user not authorized to view contract",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/model.BackendError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/echo.HTTPError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.createContractParams": {
            "type": "object",
            "required": [
                "application_id",
                "description",
                "price",
                "title"
            ],
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
        "model.BackendError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "tech_info": {
                    "type": "string"
                }
            }
        },
...
        "model.ContractDTO": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "contract_address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "customer_address": {
                    "type": "string"
                },
                "customer_display_name": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "performer_address": {
                    "type": "string"
                },
                "performer_display_name": {
                    "type": "string"
                },
                "performer_id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerToken": {
            "description": "Bearer token in Authorization header",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}